+++ src/app/app_routes.js
@@ -15,13 +15,39 @@ angular.module("gonevisDash").config(function ($stateProvider, $urlRouterProvide
       title: "Get Started"
     })
     .state("site-new", {
-      url: "/new-site?site",
+      url: "/new-blog?site",
       controller: "SiteNewController",
       template: require("./site/site_new/site_new_view.html"),
       auth: true,
-      title: "New Site",
+      title: "New Blog"
+    })
+    .state("reader", {
+      url: "/reader",
+      abstract: true,
+      template: "<ui-view/>",
+      auth: true
+    })
+    .state("reader.explore-feed", {
+      url: "/:view",
+      controller: "ReaderController",
+      template: require("./reader/reader_view.html"),
+      auth: true,
+      title: "Reader",
       params: {
-        s: 0
+        view: {
+          name: null,
+          dynamic: true
+        }
+      }
+    })
+    .state("reader.detail", {
+      url: "/post/:entryId",
+      controller: "ReaderDetailController",
+      template: require("./reader/reader_detail/reader_detail_view.html"),
+      auth: true,
+      title: "Reader",
+      params: {
+        entryId: null
       }
     })
     .state("user", {
@@ -151,10 +177,13 @@ angular.module("gonevisDash").config(function ($stateProvider, $urlRouterProvide
       }
     })
     .state("dash.comment-list", {
-      url: "/comments",
+      url: "/comments/:entryId",
       controller: "CommentController",
       template: require("./comment/comment_view.html"),
       auth: true,
+      params: {
+        entryId: null,
+      },
       title: "Comments"
     })
     .state("dash.entry-edit", {
@@ -185,6 +214,9 @@ angular.module("gonevisDash").config(function ($stateProvider, $urlRouterProvide
       controller: "EntryController",
       template: require("./entry/entry_view.html"),
       auth: true,
+      params: {
+        deletedEntry: null
+      },
       title: "Posts"
     })
     .state("dash.page-list", {
@@ -192,13 +224,28 @@ angular.module("gonevisDash").config(function ($stateProvider, $urlRouterProvide
       controller: "EntryController",
       template: require("./entry/entry_view.html"),
       auth: true,
+      params: {
+        deletedEntry: null
+      },
       title: "Pages"
     })
     .state("dash.site", {
-      url: "/site",
+      url: "/blog",
+      abstract: true,
+      template: "<ui-view/>",
+      auth: true
+    })
+    .state("dash.site.settings", {
+      url: "/:view",
       controller: "SiteController",
       template: require("./site/site_view.html"),
-      auth: true
+      auth: true,
+      params: {
+        view: {
+          name: null,
+          dynamic: true
+        }
+      }
     })
     // .state("dash.tag-edit", {
     //   url: "/tag-list/:tagId",