+++ src/app/site/index.js
@@ -1,4 +1,6 @@
+require("./payment_validation_modal/payment_validation_modal_controller");
 require("./site_controller");
 require("./site_new/site_new_controller");
 require("./site_templates_modal/site_templates_modal_controller");
 require("./site_followers/site_followers_controller");
+require("./subscription_cancellation_modal/subscription_cancellation_modal_controller");